move domain module to core
