+++ core/src/main/java/com/zizohanto/android/tobuy/core/executor/PostExecutionThread.kt
@@ -1,0 +1,9 @@
+package com.zizohanto.android.tobuy.core.executor
+
+import kotlinx.coroutines.CoroutineDispatcher
+
+interface PostExecutionThread {
+    val main: CoroutineDispatcher
+    val io: CoroutineDispatcher
+    val default: CoroutineDispatcher
+}
