+++ core/src/main/java/com/zizohanto/android/tobuy/core/models/ShoppingListWithProducts.kt
@@ -1,0 +1,9 @@
+package com.zizohanto.android.tobuy.core.models
+
+import com.zizohanto.android.tobuy.domain.sq.Product
+import com.zizohanto.android.tobuy.domain.sq.ShoppingList
+
+data class ShoppingListWithProducts(
+    val shoppingList: ShoppingList,
+    val products: List<Product>
+)
