+++ core/src/main/java/com/zizohanto/android/tobuy/core/usecase/DeleteProduct.kt
@@ -1,0 +1,19 @@
+package com.zizohanto.android.tobuy.core.usecase
+
+import com.zizohanto.android.tobuy.core.executor.PostExecutionThread
+import com.zizohanto.android.tobuy.core.repository.ProductRepository
+import com.zizohanto.android.tobuy.domain.sq.Product
+import kotlinx.coroutines.withContext
+import javax.inject.Inject
+
+class DeleteProduct @Inject constructor(
+    private val repository: ProductRepository,
+    private val postExecutionThread: PostExecutionThread
+) {
+
+    suspend operator fun invoke(product: Product) {
+        return withContext(postExecutionThread.io) {
+            repository.deleteProduct(product)
+        }
+    }
+}
