+++ core/src/main/java/com/zizohanto/android/tobuy/core/usecase/SaveShoppingList.kt
@@ -1,0 +1,19 @@
+package com.zizohanto.android.tobuy.core.usecase
+
+import com.zizohanto.android.tobuy.core.executor.PostExecutionThread
+import com.zizohanto.android.tobuy.core.repository.ShoppingListRepository
+import com.zizohanto.android.tobuy.domain.sq.ShoppingList
+import kotlinx.coroutines.withContext
+import javax.inject.Inject
+
+class SaveShoppingList @Inject constructor(
+    private val repository: ShoppingListRepository,
+    private val postExecutionThread: PostExecutionThread
+) {
+
+    suspend operator fun invoke(shoppingList: ShoppingList) {
+        withContext(postExecutionThread.io) {
+            repository.saveShoppingList(shoppingList)
+        }
+    }
+}
