+++ core/src/main/java/com/zizohanto/android/tobuy/core/utils/DateUtils.kt
@@ -1,0 +1,9 @@
+package com.zizohanto.android.tobuy.core.utils
+
+import org.joda.time.Instant
+
+object DateUtils {
+    fun getCurrentTime(): Long {
+        return Instant.now().millis
+    }
+}
