+++ core/src/main/sqldelight/com/zizohanto/android/tobuy/domain/sq/Product.sq
