--- libraries/domain/src/main/java/com/zizohanto/android/tobuy/domain/exception/NoParamsException.kt
@@ -1,25 +1,0 @@
-package com.zizohanto.android.tobuy.domain.exception
-
-import kotlin.contracts.ExperimentalContracts
-import kotlin.contracts.contract
-
-/**
- * An exception thrown when a use case parameter is null
- */
-internal class NoParamsException(errorMessage: String = noParamMessage) :
-    IllegalArgumentException(errorMessage)
-
-const val noParamMessage = "Your params cannot be null for this use case"
-
-@OptIn(ExperimentalContracts::class)
-fun <T : Any> requireParams(value: T?): T {
-    contract {
-        returns() implies (value != null)
-    }
-
-    if (value == null) {
-        throw NoParamsException()
-    } else {
-        return value
-    }
-}
