--- libraries/domain/src/main/java/com/zizohanto/android/tobuy/domain/executor/PostExecutionThread.kt
@@ -1,9 +1,0 @@
-package com.zizohanto.android.tobuy.domain.executor
-
-import kotlinx.coroutines.CoroutineDispatcher
-
-interface PostExecutionThread {
-    val main: CoroutineDispatcher
-    val io: CoroutineDispatcher
-    val default: CoroutineDispatcher
-}
