--- libraries/domain/src/main/java/com/zizohanto/android/tobuy/domain/models/ShoppingListWithProducts.kt
@@ -1,9 +1,0 @@
-package com.zizohanto.android.tobuy.domain.models
-
-import com.zizohanto.android.tobuy.domain.sq.Product
-import com.zizohanto.android.tobuy.domain.sq.ShoppingList
-
-data class ShoppingListWithProducts(
-    val shoppingList: ShoppingList,
-    val products: List<Product>
-)
