--- libraries/domain/src/main/java/com/zizohanto/android/tobuy/domain/usecase/DeleteShoppingList.kt
@@ -1,18 +1,0 @@
-package com.zizohanto.android.tobuy.domain.usecase
-
-import com.zizohanto.android.tobuy.domain.executor.PostExecutionThread
-import com.zizohanto.android.tobuy.domain.repository.ShoppingListRepository
-import kotlinx.coroutines.withContext
-import javax.inject.Inject
-
-class DeleteShoppingList @Inject constructor(
-    private val repository: ShoppingListRepository,
-    private val postExecutionThread: PostExecutionThread
-) {
-
-    suspend operator fun invoke(shoppingListId: String) {
-        withContext(postExecutionThread.io) {
-            repository.deleteShoppingList(shoppingListId)
-        }
-    }
-}
