--- libraries/domain/src/main/java/com/zizohanto/android/tobuy/domain/utils/DateUtils.kt
@@ -1,9 +1,0 @@
-package com.zizohanto.android.tobuy.domain.utils
-
-import org.joda.time.Instant
-
-object DateUtils {
-    fun getCurrentTime(): Long {
-        return Instant.now().millis
-    }
-}
