+++ shopping_list/src/main/java/com/zizohanto/android/tobuy/shopping_list/presentation/products/mvi/ProductsViewResult.kt
@@ -1,6 +1,6 @@
 package com.zizohanto.android.tobuy.shopping_list.presentation.products.mvi
 
-import com.zizohanto.android.tobuy.domain.models.ShoppingListWithProducts
+import com.zizohanto.android.tobuy.core.models.ShoppingListWithProducts
 import com.zizohanto.android.tobuy.domain.sq.Product
 import com.zizohanto.android.tobuy.domain.sq.ShoppingList
 import com.zizohanto.android.tobuy.presentation.mvi.ViewResult
