+++ shopping_list/src/main/java/com/zizohanto/android/tobuy/shopping_list/presentation/shopping_list/mvi/ShoppingListViewResult.kt
@@ -1,6 +1,6 @@
 package com.zizohanto.android.tobuy.shopping_list.presentation.shopping_list.mvi
 
-import com.zizohanto.android.tobuy.domain.models.ShoppingListWithProducts
+import com.zizohanto.android.tobuy.core.models.ShoppingListWithProducts
 import com.zizohanto.android.tobuy.domain.sq.ShoppingList
 import com.zizohanto.android.tobuy.presentation.mvi.ViewResult
